+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 ### Enhancements
 
 * Add `file_directory` to metadata
+* Add `page_name` to metadata. Currently used for the sheet name in XLSX documents.
 * Added a `--partition-strategy` parameter to unstructured-ingest so that users can specify
   partition strategy in CLI. For example, `--partition-strategy fast`.
 * Added metadata for filetype.
 * Add Discord connector to pull messages from a list of channels
+* Refactor `unstructured/file-utils/filetype.py` to better utilise hashmap to return mime type.
+* Add local declaration of DOCX_MIME_TYPES and XLSX_MIME_TYPES for `test_filetype.py`.
 
 ### Features